Use cat_file_chop_lump as transformer label
- in initializer
- currently, the table name and transformer label are overloaded.
  If the table name doesn't match the transformer label (as it did
  not in this case, where the table was cat_file_chop_lump and the
  label was cat_file_chop_lump_dunk), then the initializer does not
  find addresses associated with the _table_ name.
- temporarily setting up transformer label to match table name, but
  it would be good to circle back and decouple these things.
- also would be nice to start execute in CI to verify config works.
