+++ transformers/events/cat_file/chop_lump/transformer_test.go
@@ -14,13 +14,13 @@
 // You should have received a copy of the GNU Affero General Public License
 // along with this program.  If not, see <http://www.gnu.org/licenses/>.
 
-package chop_lump_dunk_test
+package chop_lump_test
 
 import (
 	"github.com/ethereum/go-ethereum/common"
 	"github.com/ethereum/go-ethereum/core/types"
 	"github.com/makerdao/vdb-mcd-transformers/test_config"
-	"github.com/makerdao/vdb-mcd-transformers/transformers/events/cat_file/chop_lump_dunk"
+	"github.com/makerdao/vdb-mcd-transformers/transformers/events/cat_file/chop_lump"
 	"github.com/makerdao/vdb-mcd-transformers/transformers/shared"
 	"github.com/makerdao/vdb-mcd-transformers/transformers/shared/constants"
 	"github.com/makerdao/vdb-mcd-transformers/transformers/test_data"
@@ -32,7 +32,7 @@ import (
 
 var _ = Describe("Cat file chop lump transformer", func() {
 	var (
-		transformer = chop_lump_dunk.Transformer{}
+		transformer = chop_lump.Transformer{}
 		db          = test_config.NewTestDB(test_config.NewTestNode())
 	)
